feature: integration of text analyzer
