+++ test/text_analyzer.rb
@@ -1,81 +1,43 @@
 # frozen_string_literal: true
 
-require 'open-uri'
-require 'httparty'
-require 'nokogiri'
+class TextAnalyzer
+  def analyze(text)
+    important_sentences = []
+    important_words = ["is", "are", "climate change", "net-zero", "net zero", "sustainability"]
 
-def find_companies(query)
-  options = {
-    body: {
-      company: query
-    },
-    headers: {
-      'Content-Type': 'application/x-www-form-urlencoded'
-    }
-  }
-  response = HTTParty.post('https://www.sec.gov/cgi-bin/cik_lookup', options)
-
-  return response if response.code == 200
-
-  response.body
-end
-
-def list_companies(raw_company_list)
-  companies = []
-  reports = []
-  doc = Nokogiri::HTML.parse(raw_company_list)
-  raw_companies = doc.search('a')
-  raw_companies.each do |company|
-    companies.append(company)
-  end
-  companies.delete_at(-1)
-
-  companies.each_with_index do |company, _index|
-    href = "https://www.sec.gov/edgar/#{company['href']}"
-    cik = company.text.strip
-    report_type = '10-K'
-    annual_reports = "https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=#{cik}&type=#{report_type}&dateb=&owner=include&count=40&search_text="
-    options = {
-      headers: { 'User-Agent': 'Lucas Nseyep lucas.nseyep@gmail.com' }
-    }
-    response = HTTParty.get(annual_reports, options)
-    # pp response
-    reports_html = Nokogiri::HTML.parse(response)
-
-    reports_html.search('tr').each_with_index do |report, _index|
-      reports.append(report) if report.text.strip.include?(report_type)
+    average_words = text.scan(/\w+/).count.to_f / text.scan(/(!|\.|\?)/).count
+    sentences = text.split(/(!|\.|\?)/)
+    paragraph_count = text.scan(/[^\r\n]+/).count
+    sentences.each do |sentence|
+      important_words.each do |word|
+        if sentence.scan(/\w+/).count.to_f.between?(average_words * 0.7, average_words * 1.4) && sentence =~ /\b#{word}\b/
+          important_sentences.append(sentence.strip)
+        end
+      end
     end
+    important_sentences.uniq
+  end
 
-    puts `clear`
+  private
 
-    reports.first(5).each_with_index do |report, index|
-      link = report.search('a')[1].attribute('href').value
-      date = report.search('td')[3].text
-      puts "#{index + 1}. #{date} https://www.sec.gov#{link}"
+  def open_file(file_name, array)
+    File.open(file_name, 'r').each_line do |line|
+      array.append(line)
     end
   end
-end
 
-def find_company_reports(report_type, company_cik); end
-
-def read_report
-  options = {
-    headers: { 'User-Agent': 'Lucas Nseyep lucas.nseyep@gmail.com' }
-  }
-  annual_report = 'https://www.sec.gov/cgi-bin/viewer?action=view&cik=1018724&accession_number=0001018724-23-000004&xbrl_type=v'
-  response = HTTParty.get(annual_report, options)
-  report_html = Nokogiri::HTML.parse(response)
-  path = report_html.search('#menu_cat1').at_css('a')['href']
-  href = "https://www.sec.gov/#{path.match(/Archives.+/)}"
-  response = HTTParty.get(href, options)
-  doc_html = Nokogiri::HTML.parse(response)
-  doc_html.search('span').each do |paragraph|
-    puts paragraph.text if paragraph.text.include?('climate change')
+  def most_common_words(file_name, stop_words_file_name, number_of_word)
+    words = []
+    common_word = {}
+    stop_words = []
+    open_file(file_name, words)
+    open_file(stop_words_file_name, stop_words)
+    words = words.join.gsub("\n", ' ').split
+    words.map(&:downcase!)
+    stop_words = stop_words.join.gsub("\n", ' ').split
+    words.each_with_index { |word, index| words[index] = word.match(/\w+/).to_s }
+    words -= stop_words
+    words.each { |word| common_word.include?(word) ? common_word[word] += 1 : common_word[word] = 1 }
+    common_word = common_word.sort_by { |_, count| -count }[0...number_of_word].to_h
   end
 end
-
-raw_companies = find_companies('amazon com')
-
-list_companies(raw_companies)
-
-read_report
